Move format=flowed decoding to MessageExtractor
