+++ mail/common/src/main/java/com/fsck/k9/mail/internet/FlowedMessageUtils.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.mailstore.util;
+package com.fsck.k9.mail.internet;
 
 
 /**
