+++ mail/common/src/main/java/com/fsck/k9/mail/internet/MessageExtractor.java
@@ -19,13 +19,11 @@ import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mail.Multipart;
 import com.fsck.k9.mail.Part;
-import com.fsck.k9.mail.internet.Viewable.Flowed;
 import org.apache.commons.io.input.BoundedInputStream;
 import timber.log.Timber;
 
 import static com.fsck.k9.mail.internet.CharsetSupport.fixupCharset;
 import static com.fsck.k9.mail.internet.MimeUtility.getHeaderParameter;
-import static com.fsck.k9.mail.internet.FormatFlowedHelper.isFormatFlowed;
 import static com.fsck.k9.mail.internet.MimeUtility.isSameMimeType;
 import static com.fsck.k9.mail.internet.Viewable.Alternative;
 import static com.fsck.k9.mail.internet.Viewable.Html;
@@ -112,13 +110,25 @@ public class MessageExtractor {
         InputStream in = MimeUtility.decodeBody(body);
         InputStream possiblyLimitedIn =
                 textSizeLimit != NO_TEXT_SIZE_LIMIT ? new BoundedInputStream(in, textSizeLimit) : in;
+
+        String text;
         try {
-            return CharsetSupport.readToString(possiblyLimitedIn, charset);
+            text = CharsetSupport.readToString(possiblyLimitedIn, charset);
         } finally {
             try {
                 MimeUtility.closeInputStreamWithoutDeletingTemporaryFiles(in);
             } catch (IOException e) { /* Ignore */ }
         }
+
+        if (isSameMimeType(mimeType, "text/plain")) {
+            String contentType = part.getContentType();
+            if (FormatFlowedHelper.isFormatFlowed(contentType)) {
+                boolean delSp = FormatFlowedHelper.isDelSp(contentType);
+                return FlowedMessageUtils.deflow(text, delSp);
+            }
+        }
+
+        return text;
     }
 
     public static boolean hasMissingParts(Part part) {
@@ -202,12 +212,7 @@ public class MessageExtractor {
             String mimeType = part.getMimeType();
             Viewable viewable;
             if (isSameMimeType(mimeType, "text/plain")) {
-                if (isFormatFlowed(part.getContentType())) {
-                    boolean delSp = FormatFlowedHelper.isDelSp(part.getContentType());
-                    viewable = new Flowed(part, delSp);
-                } else {
-                    viewable = new Text(part);
-                }
+                viewable = new Text(part);
             } else {
                 viewable = new Html(part);
             }
